Move AWS logic to separate file
